separate out functional code into its own module
